chore: move vanilla SearchBox stories to external file
