+++ dev/app/vanilla/stories/search-box.stories.js
@@ -1,0 +1,37 @@
+import { storiesOf } from 'dev-novel';
+import { wrapWithHits } from '../../utils/wrap-with-hits.js';
+import * as widgets from '../widgets/index.js';
+
+const stories = storiesOf('SearchBox');
+
+export default () => {
+  stories
+    .add(
+      'default',
+      wrapWithHits(container => {
+        const input = document.createElement('input');
+        container.appendChild(input);
+
+        window.search.addWidget(
+          widgets.searchBox({
+            node: input,
+            placeholder: 'Search for products',
+          })
+        );
+      })
+    )
+    .add(
+      'with enter to search',
+      wrapWithHits(container => {
+        const input = document.createElement('input');
+        container.appendChild(input);
+
+        window.search.addWidget(
+          widgets.searchBoxReturn({
+            node: input,
+            placeholder: 'Search for products',
+          })
+        );
+      })
+    );
+};
